Fix warning.

+++ src/alloy/runtime/debug_info.cc
@@ -62,7 +62,7 @@ SourceMapEntry* DebugInfo::LookupHIROffset(uint64_t offset) {
 
 SourceMapEntry* DebugInfo::LookupCodeOffset(uint64_t offset) {
   // TODO(benvanik): binary search? We know the list is sorted by code order.
-  for (int n = source_map_count_ - 1; n >= 0; n--) {
+  for (int64_t n = source_map_count_ - 1; n >= 0; n--) {
     auto entry = &source_map_[n];
     if (entry->code_offset <= offset) {
       return entry;
